Replace WpfTableControl with new HotspotsList control (#1925)
* Replace WpfTableControl with new HotspotsList control

* Rename file

* Remove Shell 15.0 references (#1926)
