--- src/IssueViz.Security.UnitTests/Helpers/ThreadHelper.cs
@@ -1,37 +1,0 @@
-﻿/*
- * SonarLint for Visual Studio
- * Copyright (C) 2016-2020 SonarSource SA
- * mailto:info AT sonarsource DOT com
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- */
-
-using FluentAssertions;
-
-namespace SonarLint.VisualStudio.IssueVisualization.Security.UnitTests.Helpers
-{
-    internal class ThreadHelper
-    {
-        /// <summary>
-        /// Shell-dependent implementation of <see cref="Integration.UnitTests.ThreadHelper"/>.
-        /// </summary>
-        public static void SetCurrentThreadAsUIThread()
-        {
-            var methodInfo = typeof(Microsoft.VisualStudio.Shell.ThreadHelper).GetMethod("SetUIThread", System.Reflection.BindingFlags.Static | System.Reflection.BindingFlags.NonPublic);
-            methodInfo.Should().NotBeNull("Could not find ThreadHelper.SetUIThread");
-            methodInfo.Invoke(null, null);
-        }
-    }
-}
