+++ src/IssueViz.Security.UnitTests/HotspotsList/HotspotsToolWindowTests.cs
@@ -19,16 +19,16 @@
  */
 
 using System;
-using System.Collections.Generic;
-using Microsoft.Internal.VisualStudio.Shell.TableControl;
+using System.Collections.ObjectModel;
 using Microsoft.VisualStudio.ComponentModelHost;
 using Microsoft.VisualStudio.Shell.Interop;
-using Microsoft.VisualStudio.Shell.TableControl;
-using Microsoft.VisualStudio.Shell.TableManager;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
+using SonarLint.VisualStudio.IssueVisualization.Editor;
+using SonarLint.VisualStudio.IssueVisualization.Models;
 using SonarLint.VisualStudio.IssueVisualization.Security.HotspotsList;
 using SonarLint.VisualStudio.IssueVisualization.Security.SelectionService;
+using SonarLint.VisualStudio.IssueVisualization.Security.Store;
 
 namespace SonarLint.VisualStudio.IssueVisualization.Security.UnitTests.HotspotsList
 {
@@ -39,7 +39,7 @@ namespace SonarLint.VisualStudio.IssueVisualization.Security.UnitTests.HotspotsL
         public void Dispose_FrameIsClosed()
         {
             var frameMock = new Mock<IVsWindowFrame>();
-            var testSubject = CreateTestSubject(frameMock: frameMock);
+            var testSubject = CreateTestSubject(frameMock);
 
             frameMock.Verify(x => x.CloseFrame(It.IsAny<uint>()), Times.Never());
 
@@ -48,42 +48,22 @@ namespace SonarLint.VisualStudio.IssueVisualization.Security.UnitTests.HotspotsL
             frameMock.Verify(x => x.CloseFrame((uint)__FRAMECLOSE.FRAMECLOSE_NoSave), Times.Once());
         }
 
-        [TestMethod]
-        public void Dispose_HotspotsControlIsDisposed()
+        private HotspotsToolWindow CreateTestSubject(Mock<IVsWindowFrame> frameMock)
         {
-            var hotspotsTableControl = new Mock<IWpfTableControl>();
-            var testSubject = CreateTestSubject(hotspotsTableControl);
-
-            hotspotsTableControl.Verify(x=> x.Dispose(), Times.Never);
-
-            testSubject.Dispose();
-
-            hotspotsTableControl.Verify(x => x.Dispose(), Times.Once);
-        }
-
-        private HotspotsToolWindow CreateTestSubject(Mock<IWpfTableControl> hotspotsTableControl = null,
-            Mock<IVsWindowFrame> frameMock = null)
-        {
-            hotspotsTableControl ??= new Mock<IWpfTableControl>();
-
-            var wpfTableControlProviderMock = new Mock<IWpfTableControlProvider>();
-            wpfTableControlProviderMock
-                .Setup(x => x.CreateControl(
-                    It.IsAny<ITableManager>(),
-                    It.IsAny<bool>(),
-                    It.IsAny<IEnumerable<ColumnState>>(),
-                    It.IsAny<string[]>()))
-                .Returns(hotspotsTableControl.Object);
-
             var componentModelMock = new Mock<IComponentModel>();
+            
+            var hotspotsStore = new Mock<IHotspotsStore>();
+            hotspotsStore.Setup(x => x.GetAll())
+                .Returns(new ReadOnlyObservableCollection<IAnalysisIssueVisualization>(
+                    new ObservableCollection<IAnalysisIssueVisualization>()));
 
             componentModelMock
-                .Setup(x => x.GetService<ITableManagerProvider>())
-                .Returns(Mock.Of<ITableManagerProvider>());
+                .Setup(x => x.GetService<IHotspotsStore>())
+                .Returns(hotspotsStore.Object);
 
             componentModelMock
-                .Setup(x => x.GetService<IWpfTableControlProvider>())
-                .Returns(wpfTableControlProviderMock.Object);
+                .Setup(x => x.GetService<ILocationNavigator>())
+                .Returns(Mock.Of<ILocationNavigator>());
 
             componentModelMock
                 .Setup(x => x.GetService<IHotspotsSelectionService>())
