+++ src/IssueViz.Security/HotspotsList/HotspotsControl.xaml.cs
@@ -18,79 +18,20 @@
  * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
  */
 
-using System;
-using System.Linq;
 using System.Windows.Controls;
-using Microsoft.Internal.VisualStudio.Shell.TableControl;
-using Microsoft.VisualStudio.Shell.TableControl;
-using Microsoft.VisualStudio.Shell.TableManager;
-using SonarLint.VisualStudio.IssueVisualization.Helpers;
-using SonarLint.VisualStudio.IssueVisualization.Security.HotspotsList.TableDataSource;
-using SonarLint.VisualStudio.IssueVisualization.Security.SelectionService;
+using SonarLint.VisualStudio.IssueVisualization.Security.HotspotsList.ViewModels;
 
 namespace SonarLint.VisualStudio.IssueVisualization.Security.HotspotsList
 {
-    internal sealed partial class HotspotsControl : UserControl, IDisposable
+    internal sealed partial class HotspotsControl : UserControl
     {
-        private readonly IHotspotsSelectionService selectionService;
-        private readonly IWpfTableControl wpfTableControl;
+        public IHotspotsControlViewModel ViewModel { get; }
 
-        public HotspotsControl(ITableManagerProvider tableManagerProvider, 
-            IWpfTableControlProvider wpfTableControlProvider,
-            IHotspotsSelectionService selectionService)
+        public HotspotsControl(IHotspotsControlViewModel viewModel)
         {
-            this.selectionService = selectionService;
+            ViewModel = viewModel;
+
             InitializeComponent();
-
-            wpfTableControl = CreateWpfTableControl(tableManagerProvider, wpfTableControlProvider);
-            hotspotsList.Child = wpfTableControl.Control;
-
-            selectionService.SelectionChanged += SelectionService_SelectionChanged;
-        }
-
-        private static IWpfTableControl CreateWpfTableControl(ITableManagerProvider tableManagerProvider, IWpfTableControlProvider wpfTableControlProvider)
-        {
-            var tableManager = tableManagerProvider.GetTableManager(HotspotsTableConstants.TableManagerIdentifier);
-
-            var wpfTableControl = wpfTableControlProvider.CreateControl(tableManager,
-                true,
-                HotspotsTableColumns.InitialStates,
-                HotspotsTableColumns.Names);
-
-            wpfTableControl.SelectionMode = SelectionMode.Single;
-
-            return wpfTableControl;
-        }
-
-        private async void SelectionService_SelectionChanged(object sender, SelectionService.SelectionChangedEventArgs e)
-        {
-            if (e.SelectedHotspot == null)
-            {
-                return;
-            }
-
-            // There is an internal delay between adding an entry to the table sink and until the WpfTableControl picks up on the changes.
-            // So it's possible that when we fire the selection event, that the entry was added to the sink but not yet picked up by the wpf control.
-            // So wpfTableControl.Entries will not contain the entry, even though it should be there.
-            // To solve it we are forcing the control to sync the entries.
-            await wpfTableControl.ForceUpdateAsync();
-
-            await RunOnUIThread.RunAsync(() =>
-            {
-                var tableEntryHandle = wpfTableControl.Entries.FirstOrDefault(x => x.Identity == e.SelectedHotspot);
-
-                if (tableEntryHandle != null)
-                {
-                    wpfTableControl.SelectedEntries = new[] {tableEntryHandle};
-                }
-            });
-        }
-
-        public void Dispose()
-        {
-            selectionService.SelectionChanged -= SelectionService_SelectionChanged;
-
-            wpfTableControl?.Dispose();
         }
     }
 }
