--- src/IssueViz.Security/HotspotsList/TableDataSource/Events/HotspotsTableEventProcessorProvider.cs
@@ -1,50 +1,0 @@
-﻿/*
- * SonarLint for Visual Studio
- * Copyright (C) 2016-2020 SonarSource SA
- * mailto:info AT sonarsource DOT com
- *
- * This program is free software; you can redistribute it and/or
- * modify it under the terms of the GNU Lesser General Public
- * License as published by the Free Software Foundation; either
- * version 3 of the License, or (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
- * Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program; if not, write to the Free Software Foundation,
- * Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
- */
-
-using System.ComponentModel.Composition;
-using Microsoft.VisualStudio.Shell.TableControl;
-using Microsoft.VisualStudio.Text.Classification;
-using Microsoft.VisualStudio.Utilities;
-using SonarLint.VisualStudio.IssueVisualization.Editor;
-
-namespace SonarLint.VisualStudio.IssueVisualization.Security.HotspotsList.TableDataSource.Events
-{
-    [Export(typeof(ITableControlEventProcessorProvider))]
-    [DataSourceType(HotspotsTableConstants.TableSourceTypeIdentifier)]
-    [DataSource(HotspotsTableConstants.TableIdentifier)]
-    [ManagerType(HotspotsTableConstants.TableManagerIdentifier)]
-    [Name(nameof(HotspotsTableEventProcessorProvider))]
-    [Order(Before = Priority.Default)]
-    internal class HotspotsTableEventProcessorProvider : ITableControlEventProcessorProvider
-    {
-        private readonly ILocationNavigator locationNavigator;
-
-        [ImportingConstructor]
-        public HotspotsTableEventProcessorProvider(ILocationNavigator locationNavigator)
-        {
-            this.locationNavigator = locationNavigator;
-        }
-
-        ITableControlEventProcessor ITableControlEventProcessorProvider.GetAssociatedEventProcessor(IWpfTableControl tableControl)
-        {
-            return new HotspotsTableEventProcessor(locationNavigator);
-        }
-    }
-}
