+++ src/IssueViz.Security/HotspotsList/ViewModels/HotspotsControlViewModel.cs
@@ -31,7 +31,7 @@ using SonarLint.VisualStudio.IssueVisualization.Models;
 using SonarLint.VisualStudio.IssueVisualization.Security.SelectionService;
 using SonarLint.VisualStudio.IssueVisualization.Security.Store;
 
-namespace SonarLint.VisualStudio.IssueVisualization.Security.HotspotsList2.ViewModels
+namespace SonarLint.VisualStudio.IssueVisualization.Security.HotspotsList.ViewModels
 {
     internal interface IHotspotsControlViewModel : IDisposable
     {
