--- src/IssueViz.Security/HotspotsList2/HotspotsControl2.xaml
@@ -1,113 +1,0 @@
-﻿<UserControl x:Class="SonarLint.VisualStudio.IssueVisualization.Security.HotspotsList2.HotspotsControl"
-             x:ClassModifier="internal"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:wpf="clr-namespace:SonarLint.VisualStudio.Core.WPF;assembly=SonarLint.VisualStudio.Core"
-             xmlns:issueVisualizationControl="clr-namespace:SonarLint.VisualStudio.IssueVisualization.IssueVisualizationControl;assembly=SonarLint.VisualStudio.IssueVisualization"
-             DataContext="{Binding ViewModel, RelativeSource={RelativeSource Mode=Self}}">
-    <UserControl.Resources>
-        <wpf:AggregatorValueConverter x:Key="SpanToInvertedVisibilityConverter">
-            <issueVisualizationControl:SpanToNavigabilityConverter />
-            <wpf:BoolToVisibilityConverter TrueValue="Collapsed" FalseValue="Visible" />
-        </wpf:AggregatorValueConverter>
-
-        <Style x:Key="CellTextBlockStyle" TargetType="TextBlock">
-            <Setter Property="TextWrapping" Value="Wrap"/>
-            <Setter Property="TextAlignment" Value="Left"/>
-        </Style>
-    </UserControl.Resources>
-    <Grid UseLayoutRounding="True">
-        <!--vsTheming:ImageThemingUtilities.ImageBackgroundColor="{Binding Background, RelativeSource={RelativeSource Self}, Converter={StaticResource BrushToColorConverter}}">-->
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Name="Table" Width="*"/>
-        </Grid.ColumnDefinitions>
-        <DataGrid Grid.Column="0" 
-                  IsReadOnly="False" 
-                  CanUserAddRows="False"
-                  CanUserDeleteRows="False"
-                  CanUserReorderColumns="True"
-                  CanUserResizeColumns="True"
-                  CanUserResizeRows="False"
-                  CanUserSortColumns="True"
-                  AutoGenerateColumns="False"
-                  SelectionMode="Single"
-                  SelectionUnit="FullRow"
-                  IsSynchronizedWithCurrentItem="True"
-                  SelectedItem="{Binding SelectedHotspot, Mode=TwoWay}"
-                  ItemsSource="{Binding Hotspots}">
-            <DataGrid.InputBindings>
-                <MouseBinding
-                    MouseAction="LeftDoubleClick"
-                    Command="{Binding NavigateCommand}"
-                    CommandParameter="{Binding Hotspots/}"/>
-                <KeyBinding
-                    Key="Enter"
-                    Command="{Binding NavigateCommand}"
-                    CommandParameter="{Binding Hotspots/}"/>
-            </DataGrid.InputBindings>
-            <DataGrid.Columns>
-                <DataGridTemplateColumn CanUserSort="False" Width="25">
-                    <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate>
-                            <Label Visibility="{Binding Hotspot.Span, Converter={StaticResource SpanToInvertedVisibilityConverter}, Mode=OneWay}">
-                                todo: image
-                            </Label>
-                            <!--<vsImaging:CrispImage Grid.Column="2"
-                                                  Height="16" Width="16"
-                                                  Moniker="{x:Static vsCatalog:KnownMonikers.DocumentWarning}" 
-                                                  Visibility="{Binding Location.Span, Converter={StaticResource SpanToInvertedVisibilityConverter}, Mode=OneWay}"/>-->
-                        </DataTemplate>
-                    </DataGridTemplateColumn.CellTemplate>
-                </DataGridTemplateColumn>
-                <DataGridTemplateColumn CanUserSort="True" SortMemberPath="Hotspot.Issue.RuleKey" Header="Code" Width="100">
-                    <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate>
-                            <TextBlock Text="{Binding Hotspot.Issue.RuleKey}" 
-                                       Style="{StaticResource CellTextBlockStyle}" />
-                        </DataTemplate>
-                    </DataGridTemplateColumn.CellTemplate>
-                </DataGridTemplateColumn>
-                <DataGridTemplateColumn CanUserSort="True" SortMemberPath="Hotspot.Issue.Priority" Header="Priority" Width="100">
-                    <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate>
-                            <TextBlock Text="{Binding Hotspot.Issue.Priority}" 
-                                       Style="{StaticResource CellTextBlockStyle}"/>
-                        </DataTemplate>
-                    </DataGridTemplateColumn.CellTemplate>
-                </DataGridTemplateColumn>
-                <DataGridTemplateColumn CanUserSort="True" SortMemberPath="Hotspot.Issue.Message" Header="Description" Width="*">
-                    <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate>
-                            <TextBlock Text="{Binding Hotspot.Issue.Message}" 
-                                       Style="{StaticResource CellTextBlockStyle}"/>
-                        </DataTemplate>
-                    </DataGridTemplateColumn.CellTemplate>
-                </DataGridTemplateColumn>
-                <DataGridTemplateColumn CanUserSort="True" SortMemberPath="FileName" Header="File" Width="200">
-                    <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate>
-                            <TextBlock Text="{Binding FileName}" 
-                                       Style="{StaticResource CellTextBlockStyle}"/>
-                        </DataTemplate>
-                    </DataGridTemplateColumn.CellTemplate>
-                </DataGridTemplateColumn>
-                <DataGridTemplateColumn CanUserSort="True" SortMemberPath="Line" Header="Line" Width="50">
-                    <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate>
-                            <TextBlock Text="{Binding Line}" 
-                                       Style="{StaticResource CellTextBlockStyle}" />
-                        </DataTemplate>
-                    </DataGridTemplateColumn.CellTemplate>
-                </DataGridTemplateColumn>
-                <DataGridTemplateColumn CanUserSort="True" SortMemberPath="Column" Header="Col" Width="50">
-                    <DataGridTemplateColumn.CellTemplate>
-                        <DataTemplate>
-                            <TextBlock Text="{Binding Column}" 
-                                       Style="{StaticResource CellTextBlockStyle}" />
-                        </DataTemplate>
-                    </DataGridTemplateColumn.CellTemplate>
-                </DataGridTemplateColumn>
-            </DataGrid.Columns>
-        </DataGrid>
-    </Grid>
-</UserControl>
