+++ src/IssueViz.Security/IssueViz.Security.csproj
@@ -11,7 +11,6 @@
   <PropertyGroup>
     <!-- Hack to prevent VS2017 from automatically attempting to upgrade the project -->
     <MinimumVisualStudioVersion>$(VisualStudioVersion)</MinimumVisualStudioVersion>
-    <RequiresVersionSpecificShell>true</RequiresVersionSpecificShell>
   </PropertyGroup>
 
   <!-- This project references the Microsoft.VisualStudio.Shell assembly which is VS-version specific -->
@@ -57,23 +56,10 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Remove="HotspotsList\HotspotsControl.xaml" />
-    <None Remove="HotspotsList2\HotspotsControl2.xaml" />
-    <Page Include="HotspotsList2\HotspotsControl2.xaml">
-      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
-      <Generator>XamlIntelliSenseFileGenerator</Generator>
-    </Page>
-    
     <Page Include="HotspotsList\HotspotsControl.xaml">
       <CopyToOutputDirectory>Never</CopyToOutputDirectory>
       <Generator>XamlIntelliSenseFileGenerator</Generator>
     </Page>
-    
-    <Compile Update="HotspotsList\HotspotsListResources.Designer.cs">
-      <DesignTime>True</DesignTime>
-      <AutoGen>True</AutoGen>
-      <DependentUpon>HotspotsListResources.resx</DependentUpon>
-    </Compile>
     
     <Compile Update="OpenInIDE\OpenInIDEResources.Designer.cs">
       <DesignTime>True</DesignTime>
@@ -86,11 +72,6 @@
       <DesignTime>True</DesignTime>
       <AutoGen>True</AutoGen>
     </Compile>
-
-    <EmbeddedResource Update="HotspotsList\HotspotsListResources.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>HotspotsListResources.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
 
     <EmbeddedResource Update="OpenInIDE\OpenInIDEResources.resx">
       <Generator>ResXFileCodeGenerator</Generator>
